move batch request handling into separate file
